Delete 2018-11-12-Snowden.md
Snowden is a difficult topic, and I'd like any potential employers to feel secure that I personally would not whistleblow.

The information at the time made Snowden out to be a hero.
Current information and news being what it is, I'm a little less certain.
But unfortunately, at this point, any damning or exonerating evidence would have to be seen through the lens of whomever benefits most from that conclusion. Sadly, this means that there can be no definitive answer as to Snowden's true intentions, in the same way that we can never know what the outcomes would have been had he handled things differently.

I would like to believe that Edward Snowden believed in his country and his actions. But I cannot know for sure.